+++ packages/blaze/domrange.js
@@ -332,9 +332,13 @@ DOMRange.prototype.destroy = function (_skipNodes) {
   DOMRange._destroy(this, _skipNodes);
 };
 
-DOMRange.prototype.containsElement = function (elem) {
+DOMRange.prototype.containsElement = function (elem, selector, event) {
+  const templateName = this.view?.name
+  ? this.view.name.split('.')[1]
+  : 'unknown template';
   if (! this.attached)
-    throw new Error("Must be attached");
+    throw new Error(`${event} event triggerd with ${selector} on ${templateName} but associated view is not be found.
+    Make sure the event doesn't destroy the view.`);
 
   // An element is contained in this DOMRange if it's possible to
   // reach it by walking parent pointers, first through the DOM and