+++ packages/blaze/view.js
@@ -489,14 +489,19 @@ Blaze._destroyView = function (view, _skipNodes) {
     return;
   view.isDestroyed = true;
 
-  Blaze._fireCallbacks(view, 'destroyed');
 
   // Destroy views and elements recursively.  If _skipNodes,
   // only recurse up to views, not elements, for the case where
   // the backend (jQuery) is recursing over the elements already.
 
-  if (view._domrange)
-    view._domrange.destroyMembers(_skipNodes);
+  if (view._domrange) view._domrange.destroyMembers(_skipNodes);
+
+  // XXX: fire callbacks after potential members are destroyed
+  // otherwise it's tracker.flush will cause the above line will
+  // not be called and their views won't be destroyed
+  // Involved issues: DOMRange "Must be attached" error, mem leak
+  
+  Blaze._fireCallbacks(view, 'destroyed');
 };
 
 Blaze._destroyNode = function (node) {
@@ -692,9 +697,11 @@ Blaze.remove = function (view) {
   while (view) {
     if (! view.isDestroyed) {
       var range = view._domrange;
-      if (range.attached && ! range.parentRange)
-        range.detach();
       range.destroy();
+
+      if (range.attached && ! range.parentRange) {
+        range.detach();
+      }
     }
 
     view = view._hasGeneratedParent && view.parentView;
@@ -888,7 +895,7 @@ Blaze._addEventMap = function (view, eventMap, thisInHandler) {
         handles.push(Blaze._EventSupport.listen(
           element, newEvents, selector,
           function (evt) {
-            if (! range.containsElement(evt.currentTarget))
+            if (! range.containsElement(evt.currentTarget, selector, newEvents))
               return null;
             var handlerThis = thisInHandler || this;
             var handlerArgs = arguments;